Update npm-publish-github-packages.yml
The failure occurred because the AWS region specified (`MY_AWS_REGION`) is not valid. 

Ensure that the `aws-region` value in your GitHub Actions workflow is set to a valid AWS region. For example, replace `MY_AWS_REGION` with a valid AWS region like `us-east-1`:

```yaml
- name: Configure AWS credentials
  uses: aws-actions/configure-aws-credentials@v1
  with:
    aws-region: us-east-1
    audience: sts.amazonaws.com
```

Please update the workflow file with the correct AWS region.